msm: kgsl: Add an input handler to power up the GPU on a touch event
On a device where the primary input is a touchscreen nobody is
surprised that touch events are often the catalyst for drawing
frames. If the GPU is in slumber we can get a jump on the action
by watching for touch events and turning the GPU on in anticipation
of a draw call. Register for events from all input drivers
returning EV_ABS events and queue a work task to power up the GPU
if it is currently in slumber.

Change-Id: Ic0dedbadbdc96ef9ddd78252ea65f8e1cdd00110
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

adapted for FLO/DEB use

Signed-off-by: Paul Reioux <reioux@gmail.com>
Signed-off-by: arter97 <qkrwngud825@gmail.com>

Conflicts:
	drivers/gpu/msm/adreno.c
	drivers/gpu/msm/adreno.h

Signed-off-by: arter97 <qkrwngud825@gmail.com>

Conflicts:
	drivers/gpu/msm/adreno_ringbuffer.c

Signed-off-by: arter97 <qkrwngud825@gmail.com>